Fix eta-pkg list command
Right now it fixes the `eta-pkg list` command which annoys me to a
bigger extent that I expected. The old ouput of `eta-pkg list` was
this:

``` shellsession
sibi::jane { ~ }-> eta-pkg list
/home/sibi/.eta/package.conf.d
   array-0.5.1.1
   base-4.8.2.0
   ghc-boot-th-8.0.1
   ghc-prim-0.4.0.0
   integer-0.5.1.0
   rts-0.1.0.0
/home/sibi/.eta/package.conf.d
   array-0.5.1.1
   base-4.8.2.0
   ghc-boot-th-8.0.1
   ghc-prim-0.4.0.0
   integer-0.5.1.0
   rts-0.1.0.0
/home/sibi/.eta/package.conf.d
   array-0.5.1.1
   base-4.8.2.0
   ghc-boot-th-8.0.1
   ghc-prim-0.4.0.0
   integer-0.5.1.0
   rts-0.1.0.0
/home/sibi/.eta/package.conf.d
   array-0.5.1.1
   base-4.8.2.0
   ghc-boot-th-8.0.1
   ghc-prim-0.4.0.0
   integer-0.5.1.0
   rts-0.1.0.0
```

Now, it gives up this:

``` shellsession
@sibi::jane { ~ }-> eta-pkg list
/home/sibi/.eta/package.conf.d
   array-0.5.1.1
   base-4.8.2.0
   ghc-prim-0.4.0.0
   integer-0.5.1.0
   rts-0.1.0.0
```

Note that I can fix this more cleanly - I can remove trying to query
user database and so on. But I'm doing it now as I feel in future we
may have a user package db for Eta. This also fixes #75.